requirements.txt
streamlit
pandas
openai
numpy
python-dotenv
